requirements.txt
pyppeteer==0.2.2
python-dotenv==0.15.0
alpaca-trade-api==0.51.0
pandas==1.1.3
colorama==0.4.4
torch==1.7.0
polygon-api-client==0.1.9
